requirements.txt
python-telegram-bot==20.3
python-dotenv==0.19.2